+++ web/src/main/java/com/homeapp/javatraining/config/JwtTokenProvider.java
@@ -7,6 +7,7 @@ import io.jsonwebtoken.JwtException;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.io.Decoders;
 import io.jsonwebtoken.security.Keys;
+import lombok.Getter;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
@@ -14,6 +15,7 @@ import javax.crypto.SecretKey;
 import java.util.Date;
 
 @Component
+@Getter
 public class JwtTokenProvider {
 
     private final SecretKey secretKey;
@@ -61,5 +63,8 @@ public class JwtTokenProvider {
         } catch (JwtException | IllegalArgumentException e) {
             return  false;
         }
+
     }
+
+
 }